pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["Danillo Guimarães <danillo.guimaraes.profissional@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.2"
django-rest-framework = "^0.1.0"
factory-boy = "^3.3.1"
django-extensions = "^3.2.3"

[tool.poetry.dev-dependencies]
pytest = "6.2.5"
factory-boy = "^3.3.1"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
